change serial on bigserial value of primary keys and referenced in columns

+++ hw06-spring_ORM_JPA/src/main/resources/db/migration/V1__initial_schema.sql
@@ -4,23 +4,23 @@
 -- drop table if exists books_authors cascade;
 
 CREATE TABLE genres (
-    id serial primary key,
+    id bigserial primary key,
     name varchar(200)
 );
 
 CREATE TABLE authors (
-    id serial primary key,
+    id bigserial primary key,
     name varchar(200)
 );
 
 CREATE TABLE books (
-    id serial primary key,
+    id bigserial primary key,
     name varchar(200) not null,
-    genre_id int references genres(id)
+    genre_id bigserial references genres(id)
 );
 
 CREATE TABLE books_authors (
-    id serial primary key,
-    book_id int references books(id) ON DELETE CASCADE,
-    author_id int references authors(id) ON DELETE CASCADE
+    id bigserial primary key,
+    book_id bigserial references books(id) ON DELETE CASCADE,
+    author_id bigserial references authors(id) ON DELETE CASCADE
 );
